fraud_detection_app/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
python-jose[cryptography]
bcrypt
pydantic
mlflow
scikit-learn
prometheus-client
prometheus-fastapi-instrumentator
google-generativeai
python-multipart          # for file uploads (future)
bcrypt==4.0.1
passlib[bcrypt]
email-validator
redis 
python-dotenv
zxcvbn

